UGENE-1210 WD: 'call variations with SAM tool' worker problem UGENE-1209 WD: UGENE crashes if 'call variations with SAM tool' worker, has empty 'Source URL' input slots
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@3429 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/plugins_3rdparty/variants/src/AssemblySamtoolsMpileup.cpp
@@ -134,6 +134,10 @@ QList<Task*> CallVariantsTask::onSubTaskFinished( Task* subTask ){
 }
 
 Task::ReportResult CallVariantsTask::report(){
+    if (!mpileupTask || mpileupTask->hasError()){
+        return ReportResult_Finished;
+    }
+    
     GUrlUtils::removeFile(mpileupTask->getBcfOutputFilePath(), stateInfo);
 
     return ReportResult_Finished;
@@ -185,6 +189,23 @@ SamtoolsMpileupTask::SamtoolsMpileupTask( const CallVariantsTaskSettings& _setti
 }
 
 void SamtoolsMpileupTask::prepare(){
+    if (settings.refSeqUrl.isEmpty()){
+        setError(tr("No reference sequence URL to do pileup"));
+        return ;
+    }
+
+    if (settings.assemblyUrls.isEmpty()){
+        setError(tr("No assembly URL to do pileup"));
+        return ;
+    }
+
+    foreach(const QString& aUrl, settings.assemblyUrls){
+        if (aUrl.isEmpty()){
+            setError(tr("There is an assembly with an empty path"));
+            return ;
+        }
+    }
+    
     //prepare tmp file
     QString tmpDirPath = AppContext::getAppSettings()->getUserAppsSettings()->getCurrentProcessTemporaryDirPath(CALL_VARIANTS_DIR);
     tmpMpileupOutputFile = GUrlUtils::prepareTmpFileLocation(tmpDirPath, "tmp", "bcf", stateInfo);
